Added zlib and lzma.
ZLIB comes as a submodule as that is maintained in a git repo. The LZMA SDK is not maintained in a git repo. However, since it is placed in the public domain, we are free to do whatever we want, hence we added a stripped version of the most recent release.